migrations: TASK-7/8/9 preflight package (FK CASCADE, Numeric(12,2), subscription unique) — NOT for main
*** This branch must NOT be merged to main without an owner maintenance
window: Railway auto-runs `alembic upgrade head` on deploy, and these are
data migrations. ***

- 0036 (TASK-7): FK + ON DELETE CASCADE on user_id for moods, budget_advice,
  notification_logs, push_tokens, ignored_alerts. Deletes orphaned rows
  first (mirrors 0022).
- 0037 (TASK-8): users.monthly_income/savings_goal and expenses.amount ->
  Numeric(12,2). Rounds >2dp values to the cent (expenses.amount was float).
- 0038 (TASK-9): partial UNIQUE(platform, original_transaction_id) on
  subscriptions. Raises (never deletes money rows) if duplicates exist.
- Model classes updated to match the post-migration schema (FK
  relationships + Numeric(12,2)); they must land WITH the migrations.
- scripts/migrations/task_7_8_9_preflight.sql: read-only production preflight
  (orphan counts, rounding counts, subscription duplicate list).
- scripts/migrations/README.md: full runbook + verification record.

Verified on real PostgreSQL 15:
- empty-DB upgrade head -> 0038 (34 tables);
- upgrade runs in ONE transaction: a 0038 block on a subscription duplicate
  rolls back 0036/0037 too (all-or-nothing, stays at 0035);
- seeded orphans deleted, live rows kept, user delete cascades to 0;
- seeded expenses.amount 12.345 -> 12.35 in Numeric(12,2);
- 0038 aborts on a duplicate and touches no subscription rows; applies after
  reconciliation;
- full downgrade 0038 -> 0035 clean.

+++ app/db/models/expense.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Date, Float, Integer, String
+from sqlalchemy import Column, Date, Integer, Numeric, String
 
 from .base import Base
 
@@ -9,5 +9,5 @@ class Expense(Base):
     id = Column(Integer, primary_key=True, index=True)
     user_id = Column(String, index=True, nullable=False)
     action = Column(String, nullable=False)
-    amount = Column(Float, nullable=False)
+    amount = Column(Numeric(12, 2), nullable=False)
     date = Column(Date, nullable=False)
